Refactor isProxyOrigin and getSourceOrigin to url.js
